Update README.md
The problem Oathentic solves
Oathentic allows users to be onboarded into web3 with the option to opt into our custodial wallet option. The main problems it solves is the authenticity of reviews in the events, activites and consumer coupons sector. Furthermore, using the Poaps to filter who purchased and who redeemed their NFTs allows us to engage and re-engage with customers with Airdrops and other rewards tailored specifically to their onchain social profile based on their interersts and social activities.